Major API changes. Added support for GHC 6.12.
sendFile' now takes a FilePath instead of a Handle, and all the
unsafeSendFile variants have been removed. This was done so that we
can provide a unified API for sendFile for all OSes and versions of
GHC.

The OS-level sendfile() call takes a file descriptor. So, ideally our
interface would be:

sendFile :: Socket -> Fd -> IO ()
sendFile' :: Socket -> Fd -> OffSet -> ByteCount -> IO ()

However, the functions for getting a Fd come from the unix library and
are not available on Windows. Previously we dealt with this by having
sendFile take a Handle, and we extract the Fd from the Handle.

But, in GHC 6.12 the 'Handle' type underwent major changes and in many
cases is no longer be backed by a file descriptor. So, using this API
could lead to runtime errors.

So, we have instead opted for the API:

sendFile :: Socket -> FilePath -> IO ()
sendFile' :: Socket -> FilePath -> OffSet -> ByteCount -> IO ()

Which is consistent across OSes and versions of GHC.